Merge branch 'master' of https://github.com/Proxmark/proxmark3

+++ client/util_posix.c
@@ -117,7 +117,7 @@ uint64_t msclock() {
 	#include <sys/timeb.h>
 	struct _timeb t;
 	_ftime(&t);
-	return 1000 * t.time + t.millitm;
+	return 1000 * (uint64_t)t.time + t.millitm;
 
 // NORMAL CODE (use _ftime_s)
 	//struct _timeb t;
